Cleaned up a bit, included original banner png.
