+++ include/gfx.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void gfxDrawSprite(gfxScreen_t screen, gfx3dSide_t side, u8* spriteData, u16 width, u16 height, s16 x, s16 y);
